Update namespacing
We had the wrong name resource, the claim/eligibility has one employment
history (singular) and it'ss entries are "employments". Fixing the
resource will make supporting deleting multiple employments (ie editing
the employment history) simpler.
