+++ app/models/policies/international_relocation_payments/eligibility.rb
@@ -10,7 +10,7 @@ module Policies
 
       attr_accessor :teacher_reference_number
 
-      attribute :employment_histories, EmploymentHistoriesType.new
+      attribute :employment_history, EmploymentHistory::Type.new
 
       def ineligible?
         false
