+++ app/views/admin/claims/employment_history/employments/new.html.erb
@@ -1,6 +1,6 @@
 <%= form_with(
   model: @form,
-  url: admin_claim_employment_histories_path(@form.claim),
+  url: admin_claim_employment_history_employments_path(@form.claim),
   builder: GOVUKDesignSystemFormBuilder::FormBuilder,
 ) do |f| %>
   <%= f.govuk_error_summary %>
