+++ config/routes.rb
@@ -139,7 +139,9 @@ Rails.application.routes.draw do
       namespace :further_education_payments do
         resources :provider_verification_emails, only: [:create]
       end
-      resources :employment_histories, only: [:new, :create], module: :claims
+      resource :employment_history, only: [], module: :claims do
+        resources :employments, only: [:new, :create], module: :employment_history
+      end
     end
 
     resources :qualification_report_uploads, only: [:new, :create]
