+++ db/migrate/20250528152757_add_employment_histories_to_irp_eligibilities.rb
@@ -2,7 +2,7 @@ class AddEmploymentHistoriesToIrpEligibilities < ActiveRecord::Migration[8.0]
   def change
     add_column(
       :international_relocation_payments_eligibilities,
-      :employment_histories,
+      :employment_history,
       :jsonb,
       default: []
     )
